@@ -4,162 +4,157 @@
   "private": true,
   "homepage": "/assets/client",
   "devDependencies": {
-    "@babel/core": "^7.8.4",
-    "@babel/plugin-proposal-class-properties": "^7.8.3",
-    "@babel/plugin-proposal-decorators": "^7.8.3",
-    "@babel/plugin-proposal-do-expressions": "^7.8.3",
-    "@babel/plugin-proposal-export-default-from": "^7.8.3",
-    "@babel/plugin-proposal-export-namespace-from": "^7.8.3",
-    "@babel/plugin-proposal-function-sent": "^7.8.3",
-    "@babel/plugin-proposal-json-strings": "^7.8.3",
-    "@babel/plugin-proposal-logical-assignment-operators": "^7.8.3",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.8.3",
-    "@babel/plugin-proposal-numeric-separator": "^7.8.3",
-    "@babel/plugin-proposal-optional-chaining": "^7.8.3",
-    "@babel/plugin-proposal-pipeline-operator": "^7.8.3",
-    "@babel/plugin-proposal-throw-expressions": "^7.8.3",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-syntax-import-meta": "^7.8.3",
-    "@babel/plugin-transform-proto-to-assign": "^7.8.3",
-    "@babel/preset-react": "^7.8.3",
-    "@babel/register": "^7.8.3",
-    "@babel/runtime": "^7.8.4",
-    "@storybook/addon-actions": "^5.3.13",
-    "@storybook/react": "^5.3.13",
-    "@wdio/browserstack-service": "^5.18.1",
-    "@wdio/cli": "^5.18.6",
-    "@wdio/local-runner": "^5.18.6",
-    "@wdio/mocha-framework": "^5.18.6",
-    "@wdio/spec-reporter": "^5.18.6",
-    "@wdio/sync": "^5.18.6",
-    "aigle": "^1.14.1",
-    "autobind-decorator": "^2.4.0",
-    "autoprefixer": "^9.7.4",
-    "ava": "^3.3.0",
-    "babel-eslint": "^10.0.3",
-    "babel-jest": "^25.1.0",
-    "babel-loader": "^8.0.6",
-    "babel-preset-react-app": "^9.1.1",
-    "cache-loader": "^4.1.0",
-    "chai": "^4.2.0",
-    "chalk": "^3.0.0",
-    "chromedriver": "^80.0.1",
-    "clean-webpack-plugin": "^3.0.0",
-    "colors": "^1.4.0",
-    "config": "^3.2.5",
-    "copy-webpack-plugin": "^5.1.1",
-    "countries-list": "^2.5.1",
-    "cross-fetch": "^3.0.4",
-    "cross-spawn": "^7.0.1",
-    "css-loader": "^3.4.2",
-    "dotenv": "^8.2.0",
-    "eslint": "^6.8.0",
-    "eslint-config-react-app": "^5.2.0",
-    "eslint-config-standard": "^14.1.0",
-    "eslint-config-standard-react": "^9.2.0",
-    "eslint-loader": "^3.0.3",
-    "eslint-plugin-flowtype": "^4.6.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.18.3",
-    "eslint-plugin-react-hooks": "^2.3.0",
-    "eslint-plugin-standard": "^4.0.1",
+    "@babel/core": "7.8.6",
+    "@babel/plugin-proposal-class-properties": "7.8.3",
+    "@babel/plugin-proposal-decorators": "7.8.3",
+    "@babel/plugin-proposal-do-expressions": "7.8.3",
+    "@babel/plugin-proposal-export-default-from": "7.8.3",
+    "@babel/plugin-proposal-export-namespace-from": "7.8.3",
+    "@babel/plugin-proposal-function-sent": "7.8.3",
+    "@babel/plugin-proposal-json-strings": "7.8.3",
+    "@babel/plugin-proposal-logical-assignment-operators": "7.8.3",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "7.8.3",
+    "@babel/plugin-proposal-numeric-separator": "7.8.3",
+    "@babel/plugin-proposal-optional-chaining": "7.8.3",
+    "@babel/plugin-proposal-pipeline-operator": "7.8.3",
+    "@babel/plugin-proposal-throw-expressions": "7.8.3",
+    "@babel/plugin-syntax-dynamic-import": "7.8.3",
+    "@babel/plugin-syntax-import-meta": "7.8.3",
+    "@babel/plugin-transform-proto-to-assign": "7.8.3",
+    "@babel/preset-react": "7.8.3",
+    "@babel/register": "7.8.6",
+    "@babel/runtime": "7.8.4",
+    "@storybook/addon-actions": "5.3.14",
+    "@storybook/react": "5.3.14",
+    "@wdio/browserstack-service": "5.18.7",
+    "@wdio/cli": "5.18.7",
+    "@wdio/local-runner": "5.18.7",
+    "@wdio/mocha-framework": "5.18.7",
+    "@wdio/spec-reporter": "5.18.7",
+    "@wdio/sync": "5.18.7",
+    "aigle": "1.14.1",
+    "autoprefixer": "9.7.4",
+    "ava": "3.5.0",
+    "babel-eslint": "10.1.0",
+    "babel-jest": "25.1.0",
+    "babel-loader": "8.0.6",
+    "babel-preset-react-app": "9.1.1",
+    "cache-loader": "4.1.0",
+    "chai": "4.2.0",
+    "chalk": "3.0.0",
+    "chromedriver": "80.0.1",
+    "clean-webpack-plugin": "3.0.0",
+    "colors": "1.4.0",
+    "config": "3.3.0",
+    "copy-webpack-plugin": "5.1.1",
+    "countries-list": "2.5.2",
+    "cross-fetch": "3.0.4",
+    "cross-spawn": "7.0.1",
+    "css-loader": "3.4.2",
+    "dotenv": "8.2.0",
+    "eslint": "6.8.0",
+    "eslint-config-react-app": "5.2.0",
+    "eslint-config-standard": "14.1.0",
+    "eslint-config-standard-react": "9.2.0",
+    "eslint-loader": "3.0.3",
+    "eslint-plugin-flowtype": "4.6.0",
+    "eslint-plugin-import": "2.20.1",
+    "eslint-plugin-jsx-a11y": "6.2.3",
+    "eslint-plugin-node": "11.0.0",
+    "eslint-plugin-promise": "4.2.1",
+    "eslint-plugin-react": "7.18.3",
+    "eslint-plugin-react-hooks": "2.5.0",
+    "eslint-plugin-standard": "4.0.1",
     "faker": "https://github.com/Marak/faker.js.git",
-    "file-loader": "^5.0.2",
-    "git-describe": "^4.0.4",
-    "graphql-cli": "^3.0.14",
-    "handlebars": "^4.7.3",
-    "handlebars-loader": "^1.7.1",
-    "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^3.2.0",
-    "i18n-iso-countries": "^4.3.1",
-    "ignore-loader": "^0.1.2",
-    "jest": "^25.1.0",
-    "js-yaml": "^3.13.1",
-    "json-loader": "^0.5.7",
-    "moment-timezone": "^0.5.27",
-    "postcss-flexbugs-fixes": "^4.2.0",
-    "postcss-loader": "^3.0.0",
-    "prettier": "^1.19.1",
-    "react-test-renderer": "^16.12.0",
-    "selenium-webdriver": "^4.0.0-alpha.5",
-    "style-loader": "^1.1.3",
-    "thread-loader": "^2.1.3",
-    "uuid": "^3.4.0",
-    "wdio-chromedriver-service": "^5.0.2",
-    "webdriverio": "^5.18.6",
-    "webpack": "^4.41.6",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3",
-    "webpack-merge": "^4.2.2",
-    "xhr2": "^0.2.0"
+    "file-loader": "5.1.0",
+    "git-describe": "4.0.4",
+    "graphql-cli": "3.0.14",
+    "handlebars": "4.7.3",
+    "handlebars-loader": "1.7.1",
+    "html-loader": "0.5.5",
+    "html-webpack-plugin": "3.2.0",
+    "i18n-iso-countries": "4.3.1",
+    "ignore-loader": "0.1.2",
+    "jest": "25.1.0",
+    "js-yaml": "3.13.1",
+    "json-loader": "0.5.7",
+    "moment-timezone": "0.5.28",
+    "postcss-flexbugs-fixes": "4.2.0",
+    "postcss-loader": "3.0.0",
+    "prettier": "1.19.1",
+    "react-test-renderer": "16.13.0",
+    "selenium-webdriver": "4.0.0-alpha.5",
+    "style-loader": "1.1.3",
+    "thread-loader": "2.1.3",
+    "wdio-chromedriver-service": "5.0.2",
+    "webdriverio": "5.18.7",
+    "webpack": "4.41.6",
+    "webpack-cli": "3.3.11",
+    "webpack-dev-server": "3.10.3",
+    "webpack-merge": "4.2.2",
+    "xhr2": "0.2.0"
   },
   "dependencies": {
-    "@apollo/react-hooks": "^3.1.3",
-    "@blueprintjs/core": "^3.23.1",
-    "@blueprintjs/datetime": "^3.15.2",
-    "@fullcalendar/core": "^4.4.0",
-    "@fullcalendar/daygrid": "^4.4.0",
-    "@fullcalendar/interaction": "^4.4.0",
-    "@fullcalendar/list": "^4.4.0",
-    "@fullcalendar/react": "^4.4.0",
-    "@fullcalendar/timegrid": "^4.4.0",
-    "apollo-boost": "^0.4.7",
-    "bootstrap": "^3.4.1",
-    "change-case": "^4.1.1",
-    "classnames": "^2.2.6",
-    "core-js": "^3.6.4",
-    "d3": "^5.15.0",
-    "draft-convert": "^2.1.8",
-    "draft-js": "^0.10.5",
-    "draft-js-buttons": "^2.0.2",
-    "draft-js-side-toolbar-plugin": "^3.0.2",
-    "draftail": "^1.3.0",
-    "email-addresses": "^3.1.0",
-    "file-saver": "^2.0.2",
-    "formik": "^2.1.4",
-    "graphiql": "^0.17.5",
-    "graphql": "^14.6.0",
-    "hopscotch": "^0.3.1",
-    "jsonpath-plus": "^3.0.0",
-    "leaflet": "^1.6.0",
-    "leaflet-defaulticon-compatibility": "^0.1.1",
-    "leaflet-geosearch": "^2.7.0",
-    "leaflet-gesture-handling": "^1.1.8",
-    "leaflet.markercluster": "^1.4.1",
-    "locale-compare-polyfill": "^0.0.2",
-    "lodash": "^4.17.15",
-    "milsymbol": "^2.0.0",
-    "moment": "^2.24.0",
-    "pluralize": "^8.0.0",
-    "prop-types": "^15.7.2",
-    "react": "^16.12.0",
-    "react-avatar-edit": "^0.8.3",
-    "react-bootstrap": "^0.33.0",
-    "react-confirm-bootstrap": "^5.3.1",
-    "react-container-dimensions": "^1.4.1",
-    "react-dom": "^16.12.0",
-    "react-mosaic-component": "^3.2.0",
-    "react-notification-badge": "^1.5.1",
-    "react-redux": "^7.1.3",
-    "react-redux-loading-bar": "^4.6.0",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.1.2",
-    "react-scroll": "^1.7.16",
-    "react-svg-text": "^0.1.2",
-    "react-tag-input": "^6.4.2",
-    "react-toastify": "^5.5.0",
-    "react-tooltip": "^4.0.3",
-    "react-ultimate-pagination": "^1.2.0",
-    "react-use-dimensions": "^1.2.1",
-    "redux": "^4.0.5",
-    "redux-persist": "^6.0.0",
-    "yup": "^0.28.1"
-  },
-  "resolutions": {
-    "autobind-decorator": "2.4.0"
+    "@apollo/react-hooks": "3.1.3",
+    "@blueprintjs/core": "3.24.0",
+    "@blueprintjs/datetime": "3.15.2",
+    "@fullcalendar/core": "4.4.0",
+    "@fullcalendar/daygrid": "4.4.0",
+    "@fullcalendar/interaction": "4.4.0",
+    "@fullcalendar/list": "4.4.0",
+    "@fullcalendar/react": "4.4.0",
+    "@fullcalendar/timegrid": "4.4.0",
+    "apollo-boost": "0.4.7",
+    "bootstrap": "3.4.1",
+    "change-case": "4.1.1",
+    "classnames": "2.2.6",
+    "core-js": "3.6.4",
+    "d3": "5.15.0",
+    "draft-convert": "2.1.8",
+    "draft-js": "0.10.5",
+    "draft-js-buttons": "2.0.2",
+    "draft-js-side-toolbar-plugin": "3.0.2",
+    "draftail": "1.3.0",
+    "email-addresses": "3.1.0",
+    "file-saver": "2.0.2",
+    "formik": "2.1.4",
+    "graphiql": "0.17.5",
+    "hopscotch": "0.3.1",
+    "jsonpath-plus": "3.0.0",
+    "leaflet": "1.6.0",
+    "leaflet-defaulticon-compatibility": "0.1.1",
+    "leaflet-geosearch": "2.7.0",
+    "leaflet-gesture-handling": "1.1.8",
+    "leaflet.markercluster": "1.4.1",
+    "locale-compare-polyfill": "0.0.2",
+    "lodash": "4.17.15",
+    "milsymbol": "2.0.0",
+    "moment": "2.24.0",
+    "pluralize": "8.0.0",
+    "prop-types": "15.7.2",
+    "react": "16.13.0",
+    "react-avatar-edit": "0.8.3",
+    "react-bootstrap": "0.33.1",
+    "react-confirm-bootstrap": "5.3.1",
+    "react-container-dimensions": "1.4.1",
+    "react-dom": "16.13.0",
+    "react-mosaic-component": "3.2.0",
+    "react-notification-badge": "1.5.1",
+    "react-redux": "7.2.0",
+    "react-redux-loading-bar": "4.6.0",
+    "react-router-bootstrap": "0.25.0",
+    "react-router-dom": "5.1.2",
+    "react-scroll": "1.7.16",
+    "react-svg-text": "0.1.2",
+    "react-tag-input": "6.4.2",
+    "react-toastify": "5.5.0",
+    "react-tooltip": "4.0.3",
+    "react-ultimate-pagination": "1.2.0",
+    "react-use-dimensions": "1.2.1",
+    "redux": "4.0.5",
+    "redux-persist": "6.0.0",
+    "use-debounce": "3.3.0",
+    "yup": "0.28.1"
   },
   "scripts": {
     "build": "NODE_ENV=production webpack --mode production --config config/webpack.client.prod.js",
