@@ -1,21 +1,19 @@
 import PropTypes from "prop-types"
 import React from "react"
 
-const Checkbox = props => {
-  return (
-    <div className="checkbox">
-      <label>
-        <input
-          className="checkbox"
-          type="checkbox"
-          checked={props.checked}
-          onChange={props.onChange}
-        />
-        {props.label}
-      </label>
-    </div>
-  )
-}
+const Checkbox = props => (
+  <div className="checkbox">
+    <label>
+      <input
+        className="checkbox"
+        type="checkbox"
+        checked={props.checked}
+        onChange={props.onChange}
+      />
+      {props.label}
+    </label>
+  </div>
+)
 Checkbox.propTypes = {
   checked: PropTypes.bool,
   onChange: PropTypes.func,
