@@ -1,5 +1,4 @@
 import { Settings } from "api"
-import autobind from "autobind-decorator"
 import { Control, CRS, Icon, Map, Marker, TileLayer } from "leaflet"
 import "leaflet-defaulticon-compatibility"
 import "leaflet-defaulticon-compatibility/dist/leaflet-defaulticon-compatibility.webpack.css"
@@ -15,11 +14,9 @@ import { MarkerClusterGroup } from "leaflet.markercluster"
 import "leaflet.markercluster/dist/MarkerCluster.css"
 import "leaflet.markercluster/dist/MarkerCluster.Default.css"
 import "leaflet/dist/leaflet.css"
-import _isEqual from "lodash/isEqual"
-import _sortBy from "lodash/sortBy"
 import { Location } from "models"
 import PropTypes from "prop-types"
-import React, { Component } from "react"
+import React, { useCallback, useEffect, useRef, useState } from "react"
 import MARKER_ICON_2X from "resources/leaflet/marker-icon-2x.png"
 import MARKER_ICON from "resources/leaflet/marker-icon.png"
 import MARKER_SHADOW from "resources/leaflet/marker-shadow.png"
@@ -60,48 +57,93 @@ const searchProvider =
   Settings.imagery.geoSearcher &&
   geoSearcherProviders[Settings.imagery.geoSearcher.provider]()
 
-export default class Leaflet extends Component {
-  static propTypes = {
-    width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-    height: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-    marginBottom: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
-    markers: PropTypes.array,
-    mapId: PropTypes.string // pass this when you have more than one map on a page
-  }
+const icon = new Icon({
+  iconUrl: MARKER_ICON,
+  iconRetinaUrl: MARKER_ICON_2X,
+  shadowUrl: MARKER_SHADOW,
+  iconSize: [25, 41],
+  iconAnchor: [12, 41],
+  popupAnchor: [1, -34],
+  tooltipAnchor: [16, -28],
+  shadowSize: [41, 41]
+})
 
-  static defaultProps = {
-    width: "100%",
-    height: "500px",
-    marginBottom: "18px"
-  }
-
-  constructor(props) {
-    super(props)
-
-    this.state = {
-      map: null,
-      center: null,
-      markerLayer: null
+const addLayers = (map, layerControl) => {
+  let defaultLayer = null
+  Settings.imagery.baseLayers.forEach(layerConfig => {
+    let layer = null
+    if (layerConfig.type === "wms") {
+      layer = new TileLayer.WMS(layerConfig.url, layerConfig.options)
+    } else if (layerConfig.type === "osm" || layerConfig.type === "tile") {
+      layer = new TileLayer(layerConfig.url, layerConfig.options)
     }
-
-    this.icon = new Icon({
-      iconUrl: MARKER_ICON,
-      iconRetinaUrl: MARKER_ICON_2X,
-      shadowUrl: MARKER_SHADOW,
-      iconSize: [25, 41],
-      iconAnchor: [12, 41],
-      popupAnchor: [1, -34],
-      tooltipAnchor: [16, -28],
-      shadowSize: [41, 41]
-    })
+    if (layer) {
+      layerControl.addBaseLayer(layer, layerConfig.name)
+    }
+    if (layerConfig.default) {
+      defaultLayer = layer
+    }
+  })
+  if (defaultLayer) {
+    map.addLayer(defaultLayer)
   }
+}
 
-  get mapId() {
-    const mapId = this.props.mapId || "default"
-    return "map-" + mapId
-  }
+const Leaflet = ({
+  width,
+  height,
+  marginBottom,
+  markers,
+  mapId: initialMapId
+}) => {
+  const mapId = "map-" + (initialMapId || "default")
+  const style = Object.assign({}, css, {
+    width: width,
+    height: height,
+    marginBottom: marginBottom
+  })
 
-  componentDidMount() {
+  const latestWidth = useRef(width)
+  const widthPropUnchanged = latestWidth.current === width
+
+  const latestHeight = useRef(height)
+  const heightPropUnchanged = latestHeight.current === height
+
+  const [map, setMap] = useState(null)
+  const [markerLayer, setMarkerLayer] = useState(null)
+  const [doInitializeMarkerLayer, setDoInitializeMarkerLayer] = useState(false)
+
+  const updateMarkerLayer = useCallback(
+    (newMarkers = []) => {
+      newMarkers.forEach(m => {
+        const latLng = Location.hasCoordinates(m)
+          ? [m.lat, m.lng]
+          : map.getCenter()
+        const marker = new Marker(latLng, {
+          icon: icon,
+          draggable: m.draggable || false,
+          autoPan: m.autoPan || false,
+          id: m.id
+        })
+        if (m.name) {
+          marker.bindPopup(m.name)
+        }
+        if (m.onMove) {
+          marker.on("move", event => m.onMove(event, map))
+        }
+        markerLayer.addLayer(marker)
+      })
+
+      if (newMarkers.length > 0) {
+        if (markerLayer.getBounds() && markerLayer.getBounds().isValid()) {
+          map.fitBounds(markerLayer.getBounds(), { maxZoom: 15 })
+        }
+      }
+    },
+    [map, markerLayer]
+  )
+
+  useEffect(() => {
     Map.addInitHook("addHandler", "gestureHandling", GestureHandling)
     const mapOptions = Object.assign(
       { zoomControl: true, gestureHandling: true },
@@ -110,124 +152,61 @@ export default class Leaflet extends Component {
         crs: CRS[Settings.imagery.mapOptions.crs]
       }
     )
-    const map = new Map(this.mapId, mapOptions).setView(
+    const newMap = new Map(mapId, mapOptions).setView(
       Settings.imagery.mapOptions.homeView.location,
       Settings.imagery.mapOptions.homeView.zoomLevel
     )
     if (searchProvider) {
-      new GeoSearchControl({ provider: searchProvider }).addTo(map)
+      new GeoSearchControl({ provider: searchProvider }).addTo(newMap)
     }
-
     const layerControl = new Control.Layers({}, {}, { collapsed: false })
+    layerControl.addTo(newMap)
+    addLayers(newMap, layerControl)
 
-    layerControl.addTo(map)
-    this.addLayers(map, layerControl)
+    setMap(newMap)
 
-    map.on("moveend", this.moveEnd)
+    const newMarkerLayer = new MarkerClusterGroup().addTo(newMap)
+    setMarkerLayer(newMarkerLayer)
 
-    const markerLayer = new MarkerClusterGroup().addTo(map)
-    this.setState({ map, markerLayer })
-  }
+    setDoInitializeMarkerLayer(true)
+  }, [mapId])
 
-  componentDidUpdate(prevProps, prevState) {
-    const prevMarkerIds = _sortBy(prevProps.markers.map(m => m.id))
-    const markerIds = _sortBy(this.props.markers.map(m => m.id))
-    if (!_isEqual(prevMarkerIds, markerIds)) {
-      const markersToAdd = this.props.markers.filter(
-        m => prevProps.markers.findIndex(pm => pm.id === m.id) === -1
-      )
-      const markersToRemove = prevProps.markers.filter(
-        pm => this.props.markers.findIndex(m => m.id === pm.id) === -1
-      )
-      this.updateMarkerLayer(markersToAdd, markersToRemove)
+  useEffect(() => {
+    if (doInitializeMarkerLayer) {
+      updateMarkerLayer(markers)
+      setDoInitializeMarkerLayer(false)
     }
+  }, [doInitializeMarkerLayer, markers, updateMarkerLayer])
 
-    if (prevState.map !== this.state.map) {
-      this.updateMarkerLayer(this.props.markers)
+  useEffect(() => {
+    if (map && !(widthPropUnchanged && heightPropUnchanged)) {
+      map.invalidateSize()
+      latestWidth.current = width
+      latestHeight.current = height
     }
-    if (
-      prevProps.width !== this.props.width ||
-      prevProps.height !== this.props.height
-    ) {
-      this.state.map.invalidateSize()
-    }
-  }
+  }, [
+    height,
+    heightPropUnchanged,
+    map,
+    markerLayer,
+    markers,
+    width,
+    widthPropUnchanged
+  ])
 
-  @autobind
-  updateMarkerLayer(markersToAdd, markersToRemove) {
-    const markers = markersToAdd || []
-    markersToRemove = markersToRemove || []
-
-    const newMarkers = []
-    const markerLayer = this.state.markerLayer
-    markers.forEach(m => {
-      const latLng = Location.hasCoordinates(m)
-        ? [m.lat, m.lng]
-        : this.state.map.getCenter()
-      const marker = new Marker(latLng, {
-        icon: this.icon,
-        draggable: m.draggable || false,
-        autoPan: m.autoPan || false,
-        id: m.id
-      })
-      if (m.name) {
-        marker.bindPopup(m.name)
-      }
-      if (m.onMove) {
-        marker.on("move", event => m.onMove(event, this.state.map))
-      }
-      newMarkers.push(marker)
-      markerLayer.addLayer(marker)
-    })
-
-    markersToRemove.forEach(m => {
-      const ml = markerLayer.getLayers().find(ml => ml.options.id === m.id)
-      markerLayer.removeLayer(ml)
-    })
-
-    if (newMarkers.length > 0) {
-      if (markerLayer.getBounds() && markerLayer.getBounds().isValid()) {
-        this.state.map.fitBounds(markerLayer.getBounds(), { maxZoom: 15 })
-      }
-    }
-  }
-
-  @autobind
-  addLayers(map, layerControl) {
-    let defaultLayer = null
-    Settings.imagery.baseLayers.forEach(layerConfig => {
-      let layer = null
-      if (layerConfig.type === "wms") {
-        layer = new TileLayer.WMS(layerConfig.url, layerConfig.options)
-      } else if (layerConfig.type === "osm" || layerConfig.type === "tile") {
-        layer = new TileLayer(layerConfig.url, layerConfig.options)
-      }
-
-      if (layer) {
-        layerControl.addBaseLayer(layer, layerConfig.name)
-      }
-      if (layerConfig.default) {
-        defaultLayer = layer
-      }
-    })
-    if (defaultLayer) {
-      map.addLayer(defaultLayer)
-    }
-  }
-
-  render() {
-    const style = Object.assign({}, css, {
-      width: this.props.width,
-      height: this.props.height,
-      marginBottom: this.props.marginBottom
-    })
-    return <div id={this.mapId} style={style} />
-  }
-
-  @autobind
-  moveEnd(event) {
-    const center = this.state.map.getCenter()
-
-    this.setState({ center: [center.lat, center.lng].join(",") })
-  }
+  return <div id={mapId} style={style} />
 }
+Leaflet.propTypes = {
+  width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  height: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  marginBottom: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  markers: PropTypes.array,
+  mapId: PropTypes.string // pass this when you have more than one map on a page
+}
+Leaflet.defaultProps = {
+  width: "100%",
+  height: "500px",
+  marginBottom: "18px"
+}
+
+export default Leaflet
