@@ -3,9 +3,10 @@ import "@blueprintjs/core/lib/css/blueprint.css"
 import { IconNames } from "@blueprintjs/icons"
 import "@blueprintjs/icons/lib/css/blueprint-icons.css" // needed for the mosaic tile buttons (expand, close)
 import classNames from "classnames"
+import _cloneDeep from "lodash/cloneDeep"
 import _dropRight from "lodash/dropRight"
 import PropTypes from "prop-types"
-import React, { Component } from "react"
+import React, { useState } from "react"
 import {
   Corner,
   createBalancedTreeFromLeaves,
@@ -20,81 +21,59 @@ import {
 import "react-mosaic-component/react-mosaic-component.css"
 import "./MosaicLayout.css"
 
-export default class MosaicLayout extends Component {
-  static propTypes = {
-    visualizations: PropTypes.arrayOf(
-      PropTypes.shape({
-        id: PropTypes.string.isRequired,
-        icons: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired, // icon names from @blueprintjs/icons IconNames
-        title: PropTypes.string.isRequired,
-        renderer: PropTypes.func.isRequired
-      })
-    ).isRequired,
-    initialNode: PropTypes.object, // FIXME: actually MosaicNode
-    description: PropTypes.string,
-    style: PropTypes.object
-  }
-
-  constructor(props) {
-    super(props)
-    this.state = {
-      currentNode: this.props.initialNode
-    }
-  }
-
-  render() {
-    return (
-      <div className="mosaic-box" style={this.props.style}>
-        <div className="mosaic-container">
-          {this.renderNavBar()}
-          <Mosaic
-            value={this.state.currentNode}
-            onChange={this.updateCurrentNode}
-            renderTile={(id, path) => {
-              const viz = this.props.visualizations.find(viz => viz.id === id)
-              return (
-                <MosaicWindow title={viz.title} path={path}>
-                  {viz.renderer(id)}
-                </MosaicWindow>
-              )
-            }}
-          />
-        </div>
+const MosaicLayout = ({ visualizations, initialNode, description, style }) => {
+  const [currentNode, setCurrentNode] = useState(initialNode)
+  return (
+    <div className="mosaic-box" style={style}>
+      <div className="mosaic-container">
+        {renderNavBar()}
+        <Mosaic
+          value={currentNode}
+          onChange={updateCurrentNode}
+          renderTile={(id, path) => {
+            const viz = visualizations.find(viz => viz.id === id)
+            return (
+              <MosaicWindow title={viz.title} path={path}>
+                {viz.renderer(id)}
+              </MosaicWindow>
+            )
+          }}
+        />
       </div>
-    )
-  }
+    </div>
+  )
 
-  renderNavBar = () => {
+  function renderNavBar() {
     return (
       <div className={classNames(Classes.NAVBAR)}>
         <div className={classNames(Classes.NAVBAR_GROUP, Classes.BUTTON_GROUP)}>
           <Icon iconSize={Icon.SIZE_LARGE} icon={IconNames.MENU} />
           <button
             className={classNames(Classes.BUTTON)}
-            onClick={this.autoArrange}
+            onClick={autoArrange}
             title="Auto Arrange"
           >
             <Icon icon={IconNames.GRID_VIEW} />
           </button>
-          {this.renderButtons()}
-          {this.props.description && (
-            <span className="chart-description">{this.props.description}</span>
+          {renderButtons()}
+          {description && (
+            <span className="chart-description">{description}</span>
           )}
         </div>
       </div>
     )
   }
 
-  renderButtons() {
+  function renderButtons() {
     const buttons = []
-    const leaves = getLeaves(this.state.currentNode)
-    this.props.visualizations.forEach(viz => {
+    const leaves = getLeaves(currentNode)
+    visualizations.forEach(viz => {
       if (!leaves.includes(viz.id)) {
         buttons.push(
           <button
             key={viz.id}
             className={classNames(Classes.BUTTON)}
-            onClick={() => this.addChart(viz.id)}
+            onClick={() => addChart(viz.id)}
             title={viz.title}
           >
             {viz.icons.map((icon, i) => (
@@ -107,23 +86,23 @@ export default class MosaicLayout extends Component {
     return buttons
   }
 
-  updateCurrentNode = currentNode => {
-    this.setState({ currentNode })
+  function updateCurrentNode(newCurrentNode) {
+    setCurrentNode(newCurrentNode)
   }
 
-  autoArrange = () => {
-    const leaves = getLeaves(this.state.currentNode)
-    this.updateCurrentNode(createBalancedTreeFromLeaves(leaves))
+  function autoArrange() {
+    const leaves = getLeaves(currentNode)
+    updateCurrentNode(createBalancedTreeFromLeaves(leaves))
   }
 
-  addChart = viz => {
-    let { currentNode } = this.state
-    if (!currentNode) {
-      currentNode = viz
+  function addChart(viz) {
+    let curNode = _cloneDeep(currentNode)
+    if (!curNode) {
+      curNode = viz
     } else {
-      const path = getPathToCorner(currentNode, Corner.TOP_RIGHT)
-      const parent = getNodeAtPath(currentNode, _dropRight(path))
-      const destination = getNodeAtPath(currentNode, path)
+      const path = getPathToCorner(curNode, Corner.TOP_RIGHT)
+      const parent = getNodeAtPath(curNode, _dropRight(path))
+      const destination = getNodeAtPath(curNode, path)
       const direction = parent ? getOtherDirection(parent.direction) : "row"
       let first
       let second
@@ -134,7 +113,7 @@ export default class MosaicLayout extends Component {
         first = viz
         second = destination
       }
-      currentNode = updateTree(currentNode, [
+      curNode = updateTree(curNode, [
         {
           path,
           spec: {
@@ -147,6 +126,21 @@ export default class MosaicLayout extends Component {
         }
       ])
     }
-    this.updateCurrentNode(currentNode)
+    updateCurrentNode(curNode)
   }
 }
+MosaicLayout.propTypes = {
+  visualizations: PropTypes.arrayOf(
+    PropTypes.shape({
+      id: PropTypes.string.isRequired,
+      icons: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired, // icon names from @blueprintjs/icons IconNames
+      title: PropTypes.string.isRequired,
+      renderer: PropTypes.func.isRequired
+    })
+  ).isRequired,
+  initialNode: PropTypes.object, // FIXME: actually MosaicNode
+  description: PropTypes.string,
+  style: PropTypes.object
+}
+
+export default MosaicLayout
