@@ -1,16 +1,14 @@
-import React, { Component } from "react"
+import React from "react"
 
 const css = {
   background: "orange"
 }
 
-export default class NoPositionBanner extends Component {
-  render() {
-    return (
-      <div className="banner" style={css}>
-        You haven't been assigned to an active position. Contact your
-        organization's super user to be added.
-      </div>
-    )
-  }
-}
+const NoPositionBanner = () => (
+  <div className="banner" style={css}>
+    You haven't been assigned to an active position. Contact your organization's
+    super user to be added.
+  </div>
+)
+
+export default NoPositionBanner
