@@ -1,64 +1,52 @@
-import autobind from "autobind-decorator"
 import "components/reactTags.css"
 import _clone from "lodash/clone"
 import PropTypes from "prop-types"
-import React, { Component } from "react"
+import React from "react"
 import { WithContext as ReactTags } from "react-tag-input"
 
-export default class ReportTags extends Component {
-  static propTypes = {
-    value: PropTypes.array,
-    suggestions: PropTypes.array,
-    onChange: PropTypes.func
-  }
+const ReportTags = ({ value, suggestions, onChange }) => {
+  return (
+    <ReactTags
+      tags={value}
+      suggestions={suggestions}
+      classNames={{
+        tag: "reportTag label label-info",
+        remove: "reportTagRemove label-info"
+      }}
+      minQueryLength={1}
+      autocomplete={1}
+      autofocus={false}
+      handleFilterSuggestions={handleTagSuggestions}
+      handleDelete={handleTagDelete}
+      handleAddition={handleTagAddition}
+    />
+  )
 
-  render() {
-    return (
-      <ReactTags
-        tags={this.props.value}
-        suggestions={this.props.suggestions}
-        classNames={{
-          tag: "reportTag label label-info",
-          remove: "reportTagRemove label-info"
-        }}
-        minQueryLength={1}
-        autocomplete={1}
-        autofocus={false}
-        handleFilterSuggestions={this.handleTagSuggestions}
-        handleDelete={this.handleTagDelete}
-        handleAddition={this.handleTagAddition}
-      />
-    )
-  }
-
-  @autobind
-  handleTagDelete(i) {
-    const tags = _clone(this.props.value)
+  function handleTagDelete(i) {
+    const tags = _clone(value)
     tags.splice(i, 1)
-    if (this.props.onChange) {
-      this.props.onChange(tags)
+    if (onChange) {
+      onChange(tags)
     }
   }
 
-  @autobind
-  handleTagAddition(tag) {
-    const newTag = this.props.suggestions.find(t => t.id === tag.id)
+  function handleTagAddition(tag) {
+    const newTag = suggestions.find(t => t.id === tag.id)
     if (newTag) {
-      const tags = _clone(this.props.value)
+      const tags = _clone(value)
       tags.push(newTag)
-      if (this.props.onChange) {
-        this.props.onChange(tags)
+      if (onChange) {
+        onChange(tags)
       }
     }
   }
 
-  @autobind
-  handleTagSuggestions(query, suggestions) {
+  function handleTagSuggestions(query, suggestions) {
     const text = (query && typeof query === "object"
       ? query.text
       : query
     ).toLowerCase()
-    const tags = this.props.value
+    const tags = value
     return suggestions.filter(
       item =>
         item.text.toLowerCase().includes(text) &&
@@ -66,3 +54,10 @@ export default class ReportTags extends Component {
     )
   }
 }
+ReportTags.propTypes = {
+  value: PropTypes.array,
+  suggestions: PropTypes.array,
+  onChange: PropTypes.func
+}
+
+export default ReportTags
