@@ -1,13 +1,26 @@
 import { SEARCH_OBJECT_LABELS, SEARCH_OBJECT_TYPES } from "actions"
 import { Settings } from "api"
-import AdvancedSelectFilter from "components/advancedSearch/AdvancedSelectFilter"
-import CheckboxSearchFilter from "components/advancedSearch/CheckboxSearchFilter"
-import DateRangeSearch from "components/advancedSearch/DateRangeSearch"
-import OrganizationFilter from "components/advancedSearch/OrganizationFilter"
-import PositionTypeSearchFilter from "components/advancedSearch/PositionTypeSearchFilter"
-import ReportStateSearch from "components/advancedSearch/ReportStateSearch"
-import SelectSearchFilter from "components/advancedSearch/SelectSearchFilter"
-import TextInputFilter from "components/advancedSearch/TextInputFilter"
+import AdvancedSelectFilter, {
+  deserialize as deserializeAdvancedSelectFilter
+} from "components/advancedSearch/AdvancedSelectFilter"
+import CheckboxFilter, {
+  deserialize as deserializeCheckboxFilter
+} from "components/advancedSearch/CheckboxFilter"
+import DateRangeFilter, {
+  deserialize as deserializeDateRangeFilter
+} from "components/advancedSearch/DateRangeFilter"
+import OrganizationFilter, {
+  deserialize as deserializeOrganizationFilter
+} from "components/advancedSearch/OrganizationFilter"
+import ReportStateFilter, {
+  deserialize as deserializeReportStateFilter
+} from "components/advancedSearch/ReportStateFilter"
+import SelectFilter, {
+  deserialize as deserializeSelectFilter
+} from "components/advancedSearch/SelectFilter"
+import TextInputFilter, {
+  deserialize as deserializeTextInputFilter
+} from "components/advancedSearch/TextInputFilter"
 import {
   LocationOverlayRow,
   PersonDetailedOverlayRow,
@@ -62,7 +75,6 @@ export const getSearchQuery = searchQuery => {
 }
 
 export const POSTITION_POSITION_TYPE_FILTER_KEY = "Position Type"
-export const POSTITION_ORGANIZATION_FILTER_KEY = "Organization"
 
 export const RECURSE_STRATEGY = {
   NONE: "NONE",
@@ -74,16 +86,18 @@ const taskFilters = () => {
   const taskFiltersObj = {
     Organization: {
       component: OrganizationFilter,
+      deserializer: deserializeOrganizationFilter,
       props: {
         queryKey: "taskedOrgUuid",
         queryOrgRecurseStrategyKey: "orgRecurseStrategy"
       }
     },
     Status: {
-      component: SelectSearchFilter,
+      component: SelectFilter,
+      deserializer: deserializeSelectFilter,
       props: {
         queryKey: "status",
-        values: [Task.STATUS.ACTIVE, Task.STATUS.INACTIVE],
+        options: [Task.STATUS.ACTIVE, Task.STATUS.INACTIVE],
         labels: ["Active", "Inactive"]
       }
     }
@@ -91,7 +105,8 @@ const taskFilters = () => {
   const projectedCompletion = Settings.fields.task.projectedCompletion
   if (projectedCompletion) {
     taskFiltersObj[projectedCompletion.label] = {
-      component: DateRangeSearch,
+      component: DateRangeFilter,
+      deserializer: deserializeDateRangeFilter,
       props: {
         queryKey: "projectedCompletion"
       }
@@ -100,7 +115,8 @@ const taskFilters = () => {
   const plannedCompletion = Settings.fields.task.plannedCompletion
   if (plannedCompletion) {
     taskFiltersObj[plannedCompletion.label] = {
-      component: DateRangeSearch,
+      component: DateRangeFilter,
+      deserializer: deserializeDateRangeFilter,
       props: {
         queryKey: "plannedCompletion"
       }
@@ -109,10 +125,11 @@ const taskFilters = () => {
   const customEnum1 = Settings.fields.task.customFieldEnum1
   if (customEnum1) {
     taskFiltersObj[customEnum1.label] = {
-      component: SelectSearchFilter,
+      component: SelectFilter,
+      deserializer: deserializeSelectFilter,
       props: {
         queryKey: "projectStatus",
-        values: Object.keys(customEnum1.enum),
+        options: Object.keys(customEnum1.enum),
         labels: Object.values(customEnum1.enum).map(o => o.label)
       }
     }
@@ -121,6 +138,7 @@ const taskFilters = () => {
   if (customField) {
     taskFiltersObj[customField.label] = {
       component: TextInputFilter,
+      deserializer: deserializeTextInputFilter,
       props: {
         queryKey: "customField"
       }
@@ -163,11 +181,12 @@ const advancedSelectFilterTaskProps = {
   addon: TASKS_ICON
 }
 
-const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
+const searchFilters = function() {
   const filters = {}
   const subscriptionFilter = {
     Subscribed: {
-      component: CheckboxSearchFilter,
+      component: CheckboxFilter,
+      deserializer: deserializeCheckboxFilter,
       props: {
         queryKey: "subscribed",
         msg: "By me"
@@ -232,6 +251,7 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
     filters: {
       Author: {
         component: AdvancedSelectFilter,
+        deserializer: deserializeAdvancedSelectFilter,
         props: Object.assign({}, advancedSelectFilterPersonProps, {
           filterDefs: authorWidgetFilters,
           placeholder: "Filter reports by author...",
@@ -240,6 +260,7 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
       },
       Attendee: {
         component: AdvancedSelectFilter,
+        deserializer: deserializeAdvancedSelectFilter,
         props: Object.assign({}, advancedSelectFilterPersonProps, {
           filterDefs: attendeeWidgetFilters,
           placeholder: "Filter reports by attendee...",
@@ -248,6 +269,7 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
       },
       "Pending Approval Of": {
         component: AdvancedSelectFilter,
+        deserializer: deserializeAdvancedSelectFilter,
         props: Object.assign({}, advancedSelectFilterPersonProps, {
           filterDefs: pendingApprovalOfWidgetFilters,
           placeholder: "Filter reports pending approval of...",
@@ -256,6 +278,7 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
       },
       "Author Position": {
         component: AdvancedSelectFilter,
+        deserializer: deserializeAdvancedSelectFilter,
         props: Object.assign({}, advancedSelectFilterPositionProps, {
           filterDefs: authorPositionWidgetFilters,
           placeholder: "Filter reports by author position...",
@@ -264,6 +287,7 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
       },
       "Attendee Position": {
         component: AdvancedSelectFilter,
+        deserializer: deserializeAdvancedSelectFilter,
         props: Object.assign({}, advancedSelectFilterPositionProps, {
           filterDefs: attendeePositionWidgetFilters,
           placeholder: "Filter reports by attendee position...",
@@ -272,37 +296,43 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
       },
       Organization: {
         component: OrganizationFilter,
+        deserializer: deserializeOrganizationFilter,
         props: {
           queryKey: "orgUuid",
           queryOrgRecurseStrategyKey: "orgRecurseStrategy"
         }
       },
       "Engagement Date": {
-        component: DateRangeSearch,
+        component: DateRangeFilter,
+        deserializer: deserializeDateRangeFilter,
         props: {
           queryKey: "engagementDate"
         }
       },
       "Release Date": {
-        component: DateRangeSearch,
+        component: DateRangeFilter,
+        deserializer: deserializeDateRangeFilter,
         props: {
           queryKey: "releasedAt"
         }
       },
       "Creation Date": {
-        component: DateRangeSearch,
+        component: DateRangeFilter,
+        deserializer: deserializeDateRangeFilter,
         props: {
           queryKey: "createdAt"
         }
       },
       "Update Date": {
-        component: DateRangeSearch,
+        component: DateRangeFilter,
+        deserializer: deserializeDateRangeFilter,
         props: {
           queryKey: "updatedAt"
         }
       },
       Location: {
         component: AdvancedSelectFilter,
+        deserializer: deserializeAdvancedSelectFilter,
         props: Object.assign({}, advancedSelectFilterLocationProps, {
           filterDefs: locationWidgetFilters,
           placeholder: "Filter reports by location...",
@@ -310,16 +340,18 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
         })
       },
       State: {
-        component: ReportStateSearch,
+        component: ReportStateFilter,
+        deserializer: deserializeReportStateFilter,
         props: {
           queryKey: "state"
         }
       },
       "Engagement Status": {
-        component: SelectSearchFilter,
+        component: SelectFilter,
+        deserializer: deserializeSelectFilter,
         props: {
           queryKey: "engagementStatus",
-          values: [
+          options: [
             Report.ENGAGEMENT_STATUS.HAPPENED,
             Report.ENGAGEMENT_STATUS.FUTURE,
             Report.ENGAGEMENT_STATUS.CANCELLED
@@ -327,14 +359,16 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
         }
       },
       [Settings.fields.report.atmosphere]: {
-        component: SelectSearchFilter,
+        component: SelectFilter,
+        deserializer: deserializeSelectFilter,
         props: {
           queryKey: "atmosphere",
-          values: ["POSITIVE", "NEUTRAL", "NEGATIVE"]
+          options: ["POSITIVE", "NEUTRAL", "NEGATIVE"]
         }
       },
       Tag: {
         component: AdvancedSelectFilter,
+        deserializer: deserializeAdvancedSelectFilter,
         props: {
           overlayColumns: ["Name"],
           overlayRenderRow: TagOverlayRow,
@@ -347,13 +381,15 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
         }
       },
       "Sensitive Info": {
-        component: CheckboxSearchFilter,
+        component: CheckboxFilter,
+        deserializer: deserializeCheckboxFilter,
         props: {
           queryKey: "sensitiveInfo"
         }
       },
       [taskShortLabel]: {
         component: AdvancedSelectFilter,
+        deserializer: deserializeAdvancedSelectFilter,
         props: Object.assign({}, advancedSelectFilterTaskProps, {
           filterDefs: taskWidgetFilters,
           placeholder: `Filter reports by ${taskShortLabel}...`,
@@ -370,16 +406,18 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
     filters: {
       Organization: {
         component: OrganizationFilter,
+        deserializer: deserializeOrganizationFilter,
         props: {
           queryKey: "orgUuid",
           queryOrgRecurseStrategyKey: "orgRecurseStrategy"
         }
       },
       Role: {
-        component: SelectSearchFilter,
+        component: SelectFilter,
+        deserializer: deserializeSelectFilter,
         props: {
           queryKey: "role",
-          values: [Person.ROLE.ADVISOR, Person.ROLE.PRINCIPAL],
+          options: [Person.ROLE.ADVISOR, Person.ROLE.PRINCIPAL],
           labels: [
             Settings.fields.advisor.person.name,
             Settings.fields.principal.person.name
@@ -387,10 +425,11 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
         }
       },
       Status: {
-        component: SelectSearchFilter,
+        component: SelectFilter,
+        deserializer: deserializeSelectFilter,
         props: {
           queryKey: "status",
-          values: [
+          options: [
             Person.STATUS.ACTIVE,
             Person.STATUS.INACTIVE,
             Person.STATUS.NEW_USER
@@ -399,6 +438,7 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
       },
       Location: {
         component: AdvancedSelectFilter,
+        deserializer: deserializeAdvancedSelectFilter,
         props: Object.assign({}, advancedSelectFilterLocationProps, {
           filterDefs: locationWidgetFilters,
           placeholder: "Filter by location...",
@@ -406,26 +446,29 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
         })
       },
       Rank: {
-        component: SelectSearchFilter,
+        component: SelectFilter,
+        deserializer: deserializeSelectFilter,
         props: {
           queryKey: "rank",
-          values: ranks,
+          options: ranks,
           labels: ranks
         }
       },
       Nationality: {
-        component: SelectSearchFilter,
+        component: SelectFilter,
+        deserializer: deserializeSelectFilter,
         props: {
           queryKey: "country",
-          values: countries,
+          options: countries,
           labels: countries
         }
       },
       "Has Biography?": {
-        component: SelectSearchFilter,
+        component: SelectFilter,
+        deserializer: deserializeSelectFilter,
         props: {
           queryKey: "hasBiography",
-          values: ["true", "false"],
+          options: ["true", "false"],
           labels: ["Yes", "No"]
         }
       },
@@ -436,17 +479,19 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
   filters[SEARCH_OBJECT_TYPES.ORGANIZATIONS] = {
     filters: {
       Status: {
-        component: SelectSearchFilter,
+        component: SelectFilter,
+        deserializer: deserializeSelectFilter,
         props: {
           queryKey: "status",
-          values: [Organization.STATUS.ACTIVE, Organization.STATUS.INACTIVE]
+          options: [Organization.STATUS.ACTIVE, Organization.STATUS.INACTIVE]
         }
       },
       "Organization Type": {
-        component: SelectSearchFilter,
+        component: SelectFilter,
+        deserializer: deserializeSelectFilter,
         props: {
           queryKey: "type",
-          values: [
+          options: [
             Organization.TYPE.ADVISOR_ORG,
             Organization.TYPE.PRINCIPAL_ORG
           ],
@@ -463,34 +508,37 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
   filters[SEARCH_OBJECT_TYPES.POSITIONS] = {
     filters: {
       [POSTITION_POSITION_TYPE_FILTER_KEY]: {
-        component: PositionTypeSearchFilter,
+        component: SelectFilter,
+        deserializer: deserializeSelectFilter,
         props: {
           queryKey: "type",
-          values: [Position.TYPE.ADVISOR, Position.TYPE.PRINCIPAL],
+          options: [Position.TYPE.ADVISOR, Position.TYPE.PRINCIPAL],
           labels: [
             Settings.fields.advisor.position.name,
             Settings.fields.principal.position.name
           ],
-          ref: positionTypeFilterRef
+          isPositionTypeFilter: true
         }
       },
-      [POSTITION_ORGANIZATION_FILTER_KEY]: {
+      Organization: {
         component: OrganizationFilter,
+        deserializer: deserializeOrganizationFilter,
         props: {
           queryKey: "organizationUuid",
-          queryOrgRecurseStrategyKey: "orgRecurseStrategy",
-          ref: organizationFilterRef
+          queryOrgRecurseStrategyKey: "orgRecurseStrategy"
         }
       },
       Status: {
-        component: SelectSearchFilter,
+        component: SelectFilter,
+        deserializer: deserializeSelectFilter,
         props: {
           queryKey: "status",
-          values: [Position.STATUS.ACTIVE, Position.STATUS.INACTIVE]
+          options: [Position.STATUS.ACTIVE, Position.STATUS.INACTIVE]
         }
       },
       Location: {
         component: AdvancedSelectFilter,
+        deserializer: deserializeAdvancedSelectFilter,
         props: Object.assign({}, advancedSelectFilterLocationProps, {
           filterDefs: locationWidgetFilters,
           placeholder: "Filter by location...",
@@ -498,10 +546,11 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
         })
       },
       "Is Filled?": {
-        component: SelectSearchFilter,
+        component: SelectFilter,
+        deserializer: deserializeSelectFilter,
         props: {
           queryKey: "isFilled",
-          values: ["true", "false"],
+          options: ["true", "false"],
           labels: ["Yes", "No"]
         }
       },
@@ -512,10 +561,11 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
   filters[SEARCH_OBJECT_TYPES.LOCATIONS] = {
     filters: {
       Status: {
-        component: SelectSearchFilter,
+        component: SelectFilter,
+        deserializer: deserializeSelectFilter,
         props: {
           queryKey: "status",
-          values: [Location.STATUS.ACTIVE, Location.STATUS.INACTIVE]
+          options: [Location.STATUS.ACTIVE, Location.STATUS.INACTIVE]
         }
       },
       ...subscriptionFilter
@@ -532,7 +582,7 @@ const searchFilters = function(positionTypeFilterRef, organizationFilterRef) {
 }
 
 // filters not being displayed in the advanced search but being used in the search
-const extraFilters = function(positionTypeFilterRef, organizationFilterRef) {
+const extraFilters = function() {
   const filters = {}
   filters[SEARCH_OBJECT_TYPES.REPORTS] = [
     "includeEngagementDayOfWeek",
@@ -570,10 +620,10 @@ SearchFilterDisplay.propTypes = {
 }
 
 export const SearchDescription = ({ searchQuery, showPlaceholders }) => {
-  const allFilters = searchFilters()
+  const ALL_FILTERS = searchFilters()
   const filterDefs =
     searchQuery.objectType && SEARCH_OBJECT_TYPES[searchQuery.objectType]
-      ? allFilters[SEARCH_OBJECT_TYPES[searchQuery.objectType]].filters
+      ? ALL_FILTERS[SEARCH_OBJECT_TYPES[searchQuery.objectType]].filters
       : {}
   const filters = searchQuery.filters.filter(f => filterDefs[f.key])
   return (
