@@ -15,12 +15,11 @@ const AdvancedSelectOverlayTable = ({
   handleRemoveItem,
   columns,
   renderRow,
-  selectItemComponent,
-  tableClassName
+  selectItemComponent
 }) => {
   const selectedItemsUuids = selectedItems.map(a => a.uuid)
   return (
-    <Table responsive hover striped className={tableClassName}>
+    <Table responsive hover striped>
       <thead>
         <tr>
           {columns.map(col => (
@@ -58,35 +57,29 @@ AdvancedSelectOverlayTable.propTypes = {
   handleRemoveItem: PropTypes.func,
   columns: PropTypes.array.isRequired,
   renderRow: PropTypes.func.isRequired,
-  selectItemComponent: PropTypes.element.isRequired,
-  tableClassName: PropTypes.string
+  selectItemComponent: PropTypes.element.isRequired
 }
 AdvancedSelectOverlayTable.defaultProps = {
-  selectedItems: [],
-  tableClassName: ""
+  selectedItems: []
 }
 
 const AdvancedSingleSelectOverlayTableBase = ({
   selectedItems,
   ...otherProps
-}) => {
-  return (
-    <AdvancedSelectOverlayTable
-      {...otherProps}
-      selectedItems={_isEmpty(selectedItems) ? [] : [selectedItems]}
-      selectItemComponent={<Radio />}
-    />
-  )
-}
+}) => (
+  <AdvancedSelectOverlayTable
+    {...otherProps}
+    selectedItems={_isEmpty(selectedItems) ? [] : [selectedItems]}
+    selectItemComponent={<Radio />}
+  />
+)
 AdvancedSingleSelectOverlayTableBase.propTypes = {
   selectedItems: PropTypes.oneOfType([PropTypes.object, PropTypes.array])
 }
 
-const AdvancedMultiSelectOverlayTableBase = props => {
-  return (
-    <AdvancedSelectOverlayTable {...props} selectItemComponent={<Checkbox />} />
-  )
-}
+const AdvancedMultiSelectOverlayTableBase = props => (
+  <AdvancedSelectOverlayTable {...props} selectItemComponent={<Checkbox />} />
+)
 
 export const AdvancedSingleSelectOverlayTable = LoaderHOC("isLoading")("items")(
   AdvancedSingleSelectOverlayTableBase
