@@ -41,146 +41,141 @@ import PropTypes from "prop-types"
 import React from "react"
 import { Redirect, Route, Switch } from "react-router-dom"
 
-const BaseRouting = ({ currentUser }) => {
-  return (
-    <Switch>
-      <Route exact path="/" component={Home} />
-      <Route path="/search" component={Search} />
-      <Route path="/rollup" component={RollupShow} />
-      <Route path="/graphiql" component={GraphiQL} />
-      <Route path="/help" component={Help} />
-      <Route
-        path="/reports"
-        render={({ match: { url } }) => (
+const BaseRouting = ({ currentUser }) => (
+  <Switch>
+    <Route exact path="/" component={Home} />
+    <Route path="/search" component={Search} />
+    <Route path="/rollup" component={RollupShow} />
+    <Route path="/graphiql" component={GraphiQL} />
+    <Route path="/help" component={Help} />
+    <Route
+      path="/reports"
+      render={({ match: { url } }) => (
+        <Switch>
+          <Route path={`${url}/new`} component={ReportNew} />
+          <Route path={`${url}/:uuid/edit`} component={ReportEdit} />
+          <Route path={`${url}/:uuid/min`} component={ReportMinimal} />
+          <Route path={`${url}/mine`} component={MyReports} />
+          <Route path={`${url}/:uuid`} component={ReportShow} />
+        </Switch>
+      )}
+    />
+    <Route
+      path="/people"
+      render={({ match: { url } }) => (
+        <Switch>
+          <Route path={`${url}/new`} component={PersonNew} />
+          <Route path={`${url}/:uuid/edit`} component={PersonEdit} />
+          <Route path={`${url}/:uuid`} component={PersonShow} />
+        </Switch>
+      )}
+    />
+    <Route
+      path="/organizations"
+      render={({ match: { url } }) => (
+        <Switch>
+          <Route path={`${url}/new`} component={OrganizationNew} />
+          <Route path={`${url}/:uuid/edit`} component={OrganizationEdit} />
+          <Route path={`${url}/:uuid/:action?`} component={OrganizationShow} />
+        </Switch>
+      )}
+    />
+    <Route
+      path="/locations"
+      render={({ match: { url } }) => (
+        <Switch>
+          <Route path={`${url}/new`} component={LocationNew} />
+          <Route path={`${url}/:uuid/edit`} component={LocationEdit} />
+          <Route path={`${url}/:uuid`} component={LocationShow} />
+        </Switch>
+      )}
+    />
+    <Route
+      path="/positions"
+      render={({ match: { url } }) => (
+        <Switch>
+          <Route path={`${url}/new`} component={PositionNew} />
+          <Route path={`${url}/:uuid/edit`} component={PositionEdit} />
+          <Route path={`${url}/:uuid`} component={PositionShow} />
+        </Switch>
+      )}
+    />
+    <Route
+      path="/tasks"
+      render={({ match: { url } }) => (
+        <Switch>
+          <Route path={`${url}/new`} component={TaskNew} />
+          <Route path={`${url}/:uuid/edit`} component={TaskEdit} />
+          <Route path={`${url}/:uuid`} component={TaskShow} />
+        </Switch>
+      )}
+    />
+    <Route
+      path="/admin"
+      render={({ match: { url } }) => (
+        <Switch>
+          <Route exact path={`${url}/`} component={AdminIndex} />
+          <Route path={`${url}/mergePeople`} component={MergePeople} />
+          <Route
+            exact
+            path={`${url}/authorizationGroups`}
+            component={AuthorizationGroups}
+          />
+          <Route
+            path={`${url}/authorizationGroups/new`}
+            component={AuthorizationGroupNew}
+          />
+          <Route
+            path={`${url}/authorizationGroups/:uuid/edit`}
+            component={AuthorizationGroupEdit}
+          />
+          <Route
+            path={`${url}/authorizationGroups/:uuid`}
+            component={AuthorizationGroupShow}
+          />
+        </Switch>
+      )}
+    />
+    <Route
+      path="/insights"
+      render={({ match: { url } }) => (
+        <Switch>
+          <Route path={`${url}/:insight`} component={InsightsShow} />
+        </Switch>
+      )}
+    />
+    <Route
+      path="/dashboards/kanban"
+      render={({ match: { url } }) => (
+        <Switch>
+          <Route path={`${url}/:dashboard`} component={KanbanDashboard} />
+        </Switch>
+      )}
+    />
+    <Route
+      path="/dashboards/decisives"
+      render={({ match: { url } }) => (
+        <Switch>
+          <Route path={`${url}/:dashboard`} component={DecisivesDashboard} />
+        </Switch>
+      )}
+    />
+    <Route
+      path="/onboarding"
+      render={({ match: { url } }) =>
+        currentUser.isNewUser() ? (
           <Switch>
-            <Route path={`${url}/new`} component={ReportNew} />
-            <Route path={`${url}/:uuid/edit`} component={ReportEdit} />
-            <Route path={`${url}/:uuid/min`} component={ReportMinimal} />
-            <Route path={`${url}/mine`} component={MyReports} />
-            <Route path={`${url}/:uuid`} component={ReportShow} />
+            <Route exact path={`${url}/`} component={OnboardingShow} />
+            <Route path={`${url}/edit`} component={OnboardingEdit} />
           </Switch>
+        ) : (
+          // Redirect to home if user account exists already. Some users bookmark the onboarding - the very first page they hit
+          <Redirect to="/" />
         )}
-      />
-      <Route
-        path="/people"
-        render={({ match: { url } }) => (
-          <Switch>
-            <Route path={`${url}/new`} component={PersonNew} />
-            <Route path={`${url}/:uuid/edit`} component={PersonEdit} />
-            <Route path={`${url}/:uuid`} component={PersonShow} />
-          </Switch>
-        )}
-      />
-      <Route
-        path="/organizations"
-        render={({ match: { url } }) => (
-          <Switch>
-            <Route path={`${url}/new`} component={OrganizationNew} />
-            <Route path={`${url}/:uuid/edit`} component={OrganizationEdit} />
-            <Route
-              path={`${url}/:uuid/:action?`}
-              component={OrganizationShow}
-            />
-          </Switch>
-        )}
-      />
-      <Route
-        path="/locations"
-        render={({ match: { url } }) => (
-          <Switch>
-            <Route path={`${url}/new`} component={LocationNew} />
-            <Route path={`${url}/:uuid/edit`} component={LocationEdit} />
-            <Route path={`${url}/:uuid`} component={LocationShow} />
-          </Switch>
-        )}
-      />
-      <Route
-        path="/positions"
-        render={({ match: { url } }) => (
-          <Switch>
-            <Route path={`${url}/new`} component={PositionNew} />
-            <Route path={`${url}/:uuid/edit`} component={PositionEdit} />
-            <Route path={`${url}/:uuid`} component={PositionShow} />
-          </Switch>
-        )}
-      />
-      <Route
-        path="/tasks"
-        render={({ match: { url } }) => (
-          <Switch>
-            <Route path={`${url}/new`} component={TaskNew} />
-            <Route path={`${url}/:uuid/edit`} component={TaskEdit} />
-            <Route path={`${url}/:uuid`} component={TaskShow} />
-          </Switch>
-        )}
-      />
-      <Route
-        path="/admin"
-        render={({ match: { url } }) => (
-          <Switch>
-            <Route exact path={`${url}/`} component={AdminIndex} />
-            <Route path={`${url}/mergePeople`} component={MergePeople} />
-            <Route
-              exact
-              path={`${url}/authorizationGroups`}
-              component={AuthorizationGroups}
-            />
-            <Route
-              path={`${url}/authorizationGroups/new`}
-              component={AuthorizationGroupNew}
-            />
-            <Route
-              path={`${url}/authorizationGroups/:uuid/edit`}
-              component={AuthorizationGroupEdit}
-            />
-            <Route
-              path={`${url}/authorizationGroups/:uuid`}
-              component={AuthorizationGroupShow}
-            />
-          </Switch>
-        )}
-      />
-      <Route
-        path="/insights"
-        render={({ match: { url } }) => (
-          <Switch>
-            <Route path={`${url}/:insight`} component={InsightsShow} />
-          </Switch>
-        )}
-      />
-      <Route
-        path="/dashboards/kanban"
-        render={({ match: { url } }) => (
-          <Switch>
-            <Route path={`${url}/:dashboard`} component={KanbanDashboard} />
-          </Switch>
-        )}
-      />
-      <Route
-        path="/dashboards/decisives"
-        render={({ match: { url } }) => (
-          <Switch>
-            <Route path={`${url}/:dashboard`} component={DecisivesDashboard} />
-          </Switch>
-        )}
-      />
-      <Route
-        path="/onboarding"
-        render={({ match: { url } }) =>
-          currentUser.isNewUser() ? (
-            <Switch>
-              <Route exact path={`${url}/`} component={OnboardingShow} />
-              <Route path={`${url}/edit`} component={OnboardingEdit} />
-            </Switch>
-          ) : (
-            // Redirect to home if user account exists already. Some users bookmark the onboarding - the very first page they hit
-            <Redirect to="/" />
-          )}
-      />
-      <Route path="*" component={PageMissing} />
-    </Switch>
-  )
-}
+    />
+    <Route path="*" component={PageMissing} />
+  </Switch>
+)
 
 BaseRouting.propTypes = {
   currentUser: PropTypes.instanceOf(Person)
