@@ -214,7 +214,7 @@ const BaseOrganizationForm = ({ currentUser, edit, title, initialValues }) => {
                   <>
                     <FastField
                       name="type"
-                      component={FieldHelper.RadioButtonToggleGroup}
+                      component={FieldHelper.RadioButtonToggleGroupField}
                       buttons={typeButtons}
                       onChange={value => setFieldValue("type", value)}
                     />
@@ -258,7 +258,7 @@ const BaseOrganizationForm = ({ currentUser, edit, title, initialValues }) => {
                     />
                     <FastField
                       name="status"
-                      component={FieldHelper.RadioButtonToggleGroup}
+                      component={FieldHelper.RadioButtonToggleGroupField}
                       buttons={statusButtons}
                       onChange={value => setFieldValue("status", value)}
                       disabled={!isAdmin}
