@@ -192,7 +192,6 @@ const BasePersonForm = ({
                 />
                 <AvatarEditModal
                   title="Edit avatar"
-                  src={currentAvatar}
                   onAvatarUpdate={onAvatarUpdate}
                 />
                 <FormGroup>
@@ -325,7 +324,7 @@ const BasePersonForm = ({
                 ) : (
                   <FastField
                     name="role"
-                    component={FieldHelper.RadioButtonToggleGroup}
+                    component={FieldHelper.RadioButtonToggleGroupField}
                     buttons={roleButtons}
                     onChange={value => {
                       const roleCountries = getCountries(value)
@@ -365,7 +364,7 @@ const BasePersonForm = ({
                 ) : (
                   <Field
                     name="status"
-                    component={FieldHelper.RadioButtonToggleGroup}
+                    component={FieldHelper.RadioButtonToggleGroupField}
                     buttons={statusButtons}
                     onChange={value => setFieldValue("status", value)}
                   >
