@@ -193,7 +193,7 @@ const BasePositionForm = ({ currentUser, edit, title, initialValues }) => {
                 ) : (
                   <FastField
                     name="type"
-                    component={FieldHelper.RadioButtonToggleGroup}
+                    component={FieldHelper.RadioButtonToggleGroupField}
                     buttons={typeButtons}
                     onChange={value => setFieldValue("type", value)}
                   />
@@ -201,7 +201,7 @@ const BasePositionForm = ({ currentUser, edit, title, initialValues }) => {
 
                 <FastField
                   name="status"
-                  component={FieldHelper.RadioButtonToggleGroup}
+                  component={FieldHelper.RadioButtonToggleGroupField}
                   buttons={statusButtons}
                   onChange={value => setFieldValue("status", value)}
                 >
@@ -258,7 +258,7 @@ const BasePositionForm = ({ currentUser, edit, title, initialValues }) => {
                 {!isPrincipal && (
                   <FastField
                     name="permissions"
-                    component={FieldHelper.RadioButtonToggleGroup}
+                    component={FieldHelper.RadioButtonToggleGroupField}
                     buttons={permissionsButtons}
                     onChange={value => setFieldValue("permissions", value)}
                   />
