@@ -7,34 +7,32 @@ const AuthorizationGroupTable = ({
   authorizationGroups,
   showDelete,
   onDelete
-}) => {
-  return (
-    <Table striped condensed hover responsive>
-      <thead>
-        <tr>
-          <th>Name</th>
-          <th>Description</th>
-          {showDelete && <th />}
+}) => (
+  <Table striped condensed hover responsive>
+    <thead>
+      <tr>
+        <th>Name</th>
+        <th>Description</th>
+        {showDelete && <th />}
+      </tr>
+    </thead>
+    <tbody>
+      {authorizationGroups.map((ag, agIndex) => (
+        <tr key={ag.uuid}>
+          <td>{ag.name}</td>
+          <td>{ag.description}</td>
+          {showDelete && (
+            <td onClick={() => onDelete(ag)}>
+              <span style={{ cursor: "pointer" }}>
+                <img src={REMOVE_ICON} height={14} alt="Remove group" />
+              </span>
+            </td>
+          )}
         </tr>
-      </thead>
-      <tbody>
-        {authorizationGroups.map((ag, agIndex) => (
-          <tr key={ag.uuid}>
-            <td>{ag.name}</td>
-            <td>{ag.description}</td>
-            {showDelete && (
-              <td onClick={() => onDelete(ag)}>
-                <span style={{ cursor: "pointer" }}>
-                  <img src={REMOVE_ICON} height={14} alt="Remove group" />
-                </span>
-              </td>
-            )}
-          </tr>
-        ))}
-      </tbody>
-    </Table>
-  )
-}
+      ))}
+    </tbody>
+  </Table>
+)
 AuthorizationGroupTable.propTypes = {
   authorizationGroups: PropTypes.array.isRequired,
   showDelete: PropTypes.bool,
