@@ -196,7 +196,7 @@ const BaseTaskForm = ({ currentUser, edit, title, initialValues }) => {
                 ) : (
                   <FastField
                     name="status"
-                    component={FieldHelper.RadioButtonToggleGroup}
+                    component={FieldHelper.RadioButtonToggleGroupField}
                     buttons={statusButtons}
                     onChange={value => setFieldValue("status", value)}
                   />
@@ -340,7 +340,7 @@ const BaseTaskForm = ({ currentUser, edit, title, initialValues }) => {
                       component={
                         disabled
                           ? FieldHelper.ReadonlyField
-                          : FieldHelper.RadioButtonToggleGroup
+                          : FieldHelper.RadioButtonToggleGroupField
                       }
                       buttons={FieldHelper.customEnumButtons(
                         Settings.fields.task.customFieldEnum1.enum
@@ -383,7 +383,7 @@ const BaseTaskForm = ({ currentUser, edit, title, initialValues }) => {
                     component={
                       disabled
                         ? FieldHelper.ReadonlyField
-                        : FieldHelper.RadioButtonToggleGroup
+                        : FieldHelper.RadioButtonToggleGroupField
                     }
                     buttons={FieldHelper.customEnumButtons(
                       Settings.fields.task.customFieldEnum2.enum
